fix(schedule): separate kv storage adapter
