+++ packages/schedule/src/runtime/kv.ts
@@ -1,7 +1,7 @@
 import * as kvPackage from "@vite-hub/kv"
 
-import { createScheduleKVStorage } from "./store.ts"
+import { createScheduleKVStorage } from "./kv-storage.ts"
 
-import type { ScheduleKVStorage } from "./store.ts"
+import type { ScheduleKVStorage } from "./kv-storage.ts"
 
 export const scheduleKVStorage: ScheduleKVStorage = createScheduleKVStorage(kvPackage.kv)
